requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crcmod==1.7
cryptography==42.0.8
decorator==5.1.1
distro==1.9.0
Django==5.1.1
dnspython==2.6.1
google==3.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jmespath==0.10.0
notion-client==2.0.0
numpy==2.0.1
openai==1.47.1
oss2==2.18.2
pandas==2.2.2
protobuf==5.28.2
py==1.11.0
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2020.5
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlparse==0.5.1
tabulate==0.9.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.6
volcengine==1.0.154
volcengine-python-sdk==1.0.102
